proxy: reduce memory usage of incoming requests
When an incoming request do not need a full-size buffer (most of the
time), copy it to a smaller buffer using the previously introduced
multisized sync pools. We trade more memory effiency by some CPU
overhead.

For TCP, there would have been possible to do multiple reads instead,
but at first, we just apply the same solution than UDP. This choice
can be revisited later.

Also, to avoid reallocation of the slice header, we ensure we do not
copy the slice we got from the pool.